+++ agent/src/main/java/com/teamscale/jacoco/agent/testimpact/TestwiseCoverageAgent.java
@@ -7,26 +7,33 @@ package com.teamscale.jacoco.agent.testimpact;
 
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Moshi;
+import com.squareup.moshi.Types;
+import com.teamscale.client.ClusteredTestDetails;
+import com.teamscale.client.StringUtils;
 import com.teamscale.client.TeamscaleServer;
 import com.teamscale.jacoco.agent.AgentBase;
 import com.teamscale.jacoco.agent.JacocoRuntimeController.DumpException;
 import com.teamscale.jacoco.agent.options.AgentOptions;
+import com.teamscale.report.testwise.jacoco.JaCoCoTestwiseReportGenerator;
 import com.teamscale.report.testwise.jacoco.cache.CoverageGenerationException;
 import com.teamscale.report.testwise.model.RevisionInfo;
 import com.teamscale.report.testwise.model.TestExecution;
-
 import spark.Request;
 import spark.Response;
+import spark.Service;
 
 import java.io.IOException;
+import java.io.PrintWriter;
+import java.io.StringWriter;
+import java.util.Collections;
+import java.util.List;
 import java.util.Optional;
 
 import static javax.servlet.http.HttpServletResponse.SC_BAD_REQUEST;
+import static javax.servlet.http.HttpServletResponse.SC_INTERNAL_SERVER_ERROR;
 import static javax.servlet.http.HttpServletResponse.SC_NO_CONTENT;
 import static javax.servlet.http.HttpServletResponse.SC_OK;
 import static org.eclipse.jetty.http.MimeTypes.Type.APPLICATION_JSON;
-import static spark.Spark.get;
-import static spark.Spark.post;
 
 /**
  * A wrapper around the JaCoCo Java agent that starts a HTTP server and listens for test events.
@@ -43,26 +50,92 @@ public class TestwiseCoverageAgent extends AgentBase {
 	/** JSON adapter for revision information. */
 	private final JsonAdapter<RevisionInfo> revisionInfoJsonAdapter = new Moshi.Builder().build()
 			.adapter(RevisionInfo.class);
-	
+
+	/** JSON adapter for test details. */
+	private final JsonAdapter<List<ClusteredTestDetails>> clusteredTestDetailsAdapter = new Moshi.Builder().build()
+			.adapter(Types.newParameterizedType(List.class, ClusteredTestDetails.class));
+
 	private final TestEventHandlerStrategyBase testEventHandler;
 
-	/** Constructor. */
-	public TestwiseCoverageAgent(AgentOptions options,
-								 TestExecutionWriter testExecutionWriter) throws IllegalStateException, CoverageGenerationException {
+	public TestwiseCoverageAgent(AgentOptions options, TestExecutionWriter testExecutionWriter,
+								 JaCoCoTestwiseReportGenerator reportGenerator) throws IllegalStateException {
 		super(options);
-		if (options.shouldDumpCoverageViaHttp()) {
-			testEventHandler = new CoverageViaHttpStrategy(options, controller);
-		} else {
-			testEventHandler = new CoverageToExecFileStrategy(testExecutionWriter, controller);
+
+		switch (options.getTestWiseCoverageMode()) {
+			case TEAMSCALE_REPORT:
+				testEventHandler = new CoverageToTeamscaleStrategy(controller, options, reportGenerator);
+				break;
+			case HTTP:
+				testEventHandler = new CoverageViaHttpStrategy(controller, options, reportGenerator);
+				break;
+			default:
+				testEventHandler = new CoverageToExecFileStrategy(controller, options, testExecutionWriter);
+				break;
 		}
 	}
 
 	@Override
-	protected void initServerEndpoints() {
-		get("/test", (request, response) -> controller.getSessionId());
-		get("/revision", (request, response) -> this.getRevisionInfo());
-		post("/test/start/" + TEST_ID_PARAMETER, this::handleTestStart);
-		post("/test/end/" + TEST_ID_PARAMETER, this::handleTestEnd);
+	protected void initServerEndpoints(Service spark) {
+		spark.get("/test", (request, response) -> controller.getSessionId());
+		spark.get("/revision", (request, response) -> this.getRevisionInfo());
+		spark.post("/test/start/" + TEST_ID_PARAMETER, this::handleTestStart);
+		spark.post("/test/end/" + TEST_ID_PARAMETER, this::handleTestEnd);
+		spark.post("/testrun/start", this::handleTestRunStart);
+		spark.post("/testrun/end", this::handleTestRunEnd);
+		spark.exception(Exception.class, this::handleThrowable);
+	}
+
+	private void handleThrowable(Exception exception, Request request, Response response) {
+		logger.error("Request to {} failed with an exception", request.pathInfo(), exception);
+
+		// we want to print stack traces to make it easier to debug problems in the agent. Otherwise, end users
+		// only see "500 - Internal server error". This is especially cumbersome when talking to the agent through
+		// the TIA Java library
+		response.status(SC_INTERNAL_SERVER_ERROR);
+		StringWriter stringWriter = new StringWriter();
+		try (PrintWriter printWriter = new PrintWriter(stringWriter)) {
+			exception.printStackTrace(printWriter);
+		}
+		response.body("Request failed with an exception in the agent: " + exception.getMessage() + "\n" + stringWriter
+				.toString());
+	}
+
+	private String handleTestRunStart(Request request, Response response) throws IOException {
+		boolean includeNonImpactedTests = "true".equalsIgnoreCase(request.params("include-non-impacted"));
+
+		String baselineParameter = request.params("baseline");
+		Long baseline = null;
+		if (baselineParameter != null) {
+			try {
+				baseline = Long.parseLong(baselineParameter);
+			} catch (NumberFormatException e) {
+				logger.error("Invalid value for parameter 'baseline'", e);
+				response.status(SC_BAD_REQUEST);
+				return "Invalid value for parameter 'baseline': " + e.getMessage();
+			}
+		}
+
+		String bodyString = request.body();
+		List<ClusteredTestDetails> availableTests = Collections.emptyList();
+		if (!StringUtils.isEmpty(bodyString)) {
+			try {
+				availableTests = clusteredTestDetailsAdapter.fromJson(bodyString);
+			} catch (IOException e) {
+				logger.error("Invalid request body. Expected a JSON list of ClusteredTestDetails", e);
+				response.status(SC_BAD_REQUEST);
+				return "Invalid request body. Expected a JSON list of ClusteredTestDetails: " + e.getMessage();
+			}
+		}
+
+		String responseBody = testEventHandler.testRunStart(availableTests, includeNonImpactedTests, baseline);
+		response.type(APPLICATION_JSON.asString());
+		return responseBody;
+	}
+
+	private String handleTestRunEnd(Request request, Response response) throws IOException {
+		testEventHandler.testRunEnd();
+		response.status(SC_NO_CONTENT);
+		return "";
 	}
 
 	/** Handles the start of a new test case by setting the session ID. */
@@ -83,7 +156,7 @@ public class TestwiseCoverageAgent extends AgentBase {
 	}
 
 	/** Handles the end of a test case by resetting the session ID. */
-	private String handleTestEnd(Request request, Response response) throws DumpException {
+	private String handleTestEnd(Request request, Response response) throws DumpException, CoverageGenerationException {
 		String testId = request.params(TEST_ID_PARAMETER);
 		if (testId == null || testId.isEmpty()) {
 			logger.error("Test name missing in " + request.url() + "!");
@@ -124,7 +197,7 @@ public class TestwiseCoverageAgent extends AgentBase {
 			return Optional.empty();
 		}
 	}
-	
+
 	/** Returns revision information for the Teamscale upload. */
 	private String getRevisionInfo() {
 		TeamscaleServer server = options.getTeamscaleServerOptions();