+++ agent/src/test/java/com/teamscale/jacoco/agent/convert/ConverterTest.java
@@ -1,7 +1,6 @@
 package com.teamscale.jacoco.agent.convert;
 
 import org.conqat.lib.commons.filesystem.FileSystemUtils;
-import org.conqat.lib.commons.test.ManagedTestData;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
@@ -13,15 +12,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 /** Basic smoke test for the converter. */
 public class ConverterTest {
 
-	private ManagedTestData managedTestData = new ManagedTestData(getClass());
-
 	/**
 	 * Ensures that running the converter on valid input does not yield any errors and produces a coverage XML report.
 	 */
 	@Test
 	public void testSmokeTest(@TempDir File tempDir) throws Exception {
-		File execFile = managedTestData.useDataFile("coverage.exec");
-		File classFile = managedTestData.useDataFile("TestClass.class");
+		File execFile = new File(getClass().getResource("coverage.exec").toURI());
+		File classFile = new File(getClass().getResource("TestClass.class").toURI());
 		File outputFile = new File(tempDir, "coverage.xml");
 
 		ConvertCommand arguments = new ConvertCommand();